fix jest using gatsby-link
